bitcoin merge

+++ Jenkinsfile
@@ -13,7 +13,6 @@ pipeline {
     }
     
     stages {
-        
         stage ('Clone Repo') {
              steps {
                 ansiColor('xterm') {
@@ -62,10 +61,12 @@ pipeline {
             steps {
                 ansiColor('xterm') {
                     // re-run the linter and save the report file
-                    sh 'cd $WORKSPACE/.docker && docker exec $(docker-compose ps -q bo) sh -c "cd /backoffice && php vendor/bin/phplint src > phplintreport.txt"'
+                    sh 'cd $WORKSPACE/.docker && docker exec $(docker-compose ps -q bo) sh -c "cd /backoffice && php vendor/bin/phplint src > phplintreport.txt || true"'
                     // run linter
                     // we need to re run if because saving the file causes a "silent" exit, running it plainly will cause the build to fail if syntax errors are found
                     sh 'cd $WORKSPACE/.docker && docker exec $(docker-compose ps -q bo) sh -c "cd /backoffice && php vendor/bin/phplint src"'
+
+                    sh 'cd $WORKSPACE/.docker && docker exec $(docker-compose ps -q bo) sh -c "cd /backoffice && (php vendor/bin/phpstan analyse src > phpstanreport.txt || true )"'
                 }
             }
         }
@@ -75,14 +76,18 @@ pipeline {
             steps {
                 ansiColor('xterm') {
                     // run PHPCS tests and create report files
-                    sh 'cd $WORKSPACE/.docker && (docker exec $(docker-compose ps -q bo) bin/sh -c "cd /backoffice; vendor/bin/phpcs --standard=PSR2 --report=full --ignore=vendor,themes,testsOld,docs,web,var/cache,build ." || true) > ../psr2report.txt'
-                    sh 'cd $WORKSPACE/.docker && (docker exec $(docker-compose ps -q bo) bin/sh -c "cd /backoffice; vendor/bin/phpcs --standard=PSR2 --report=source --ignore=vendor,themes,testsOld,docs,web,var/cache,build ." || true) > ../psr2violations_per_type.txt'
+                    sh 'cd $WORKSPACE/.docker && (docker exec $(docker-compose ps -q bo) bin/sh -c "cd /backoffice; vendor/bin/phpcs --standard=PSR2 --report=full --ignore=vendor,themes,testsOld,docs,web,var/cache,build ." || true) > ../psr2report.txt || true'
+                    sh 'cd $WORKSPACE/.docker && (docker exec $(docker-compose ps -q bo) bin/sh -c "cd /backoffice; vendor/bin/phpcs --standard=PSR2 --report=source --ignore=vendor,themes,testsOld,docs,web,var/cache,build ." || true) > ../psr2violations_per_type.txt || true'
+                    sh 'cd $WORKSPACE/.docker && (docker exec $(docker-compose ps -q bo) bin/sh -c "cd /backoffice; vendor/bin/phpcs --standard=PSR2 --report=full --ignore=vendor,themes,testsOld,docs,web,var/cache,build ." || true)'
                     sh 'cd $WORKSPACE/.docker && (docker exec $(docker-compose ps -q bo) bin/sh -c "cd /backoffice; vendor/bin/phpcs --standard=PSR2 --report=checkstyle --ignore=vendor,themes,testsOld,docs,web,var/cache,build ." || true) > ../checkstyle.xml'
 
                     // install xdebug (required by unit test coverage report)
                     sh 'cd $WORKSPACE/.docker && docker exec $(docker-compose ps -q bo) sh -c "apk add --no-cache --repository http://dl-3.alpinelinux.org/alpine/edge/testing vips-tools vips-dev fftw-dev glib-dev php7-dev php7-pear build-base && pecl install xdebug && echo \'zend_extension=/usr/lib/php7/modules/xdebug.so\' >> /etc/php7/php.ini && php -m | grep xdebug"'
-                    
-                    sh 'cd $WORKSPACE/.docker && docker exec $(docker-compose ps -q bo) sh -c "cd /backoffice && export SYMFONY_DEPRECATIONS_HELPER=disabled &&  php vendor/bin/codecept run acceptance,unit,integration,webapi --coverage-html --html -vvv"'
+
+                    // disabled generating code coverage reports for now due to "no code coverage driver available" even if XDebug is installed
+                    //sh 'cd $WORKSPACE/.docker && docker exec $(docker-compose ps -q bo) sh -c "cd /backoffice && export SYMFONY_DEPRECATIONS_HELPER=disabled && chmod 777 var -Rf && composer install && php app/console cache:clear --env=test && chmod 777 var -Rf && mkdir -p tests/_output && chmod 777 tests/_output -Rf && php vendor/bin/codecept run acceptance,unit,integration,webapi --coverage-html --html -vvv"'
+                    // run tests without code coverage
+                    sh 'cd $WORKSPACE/.docker && docker exec $(docker-compose ps -q bo) sh -c "cd /backoffice && export SYMFONY_DEPRECATIONS_HELPER=disabled && chmod 777 var -Rf && composer install && php app/console cache:clear --env=test && chmod 777 var -Rf && mkdir -p tests/_output && chmod 777 tests/_output -Rf && php vendor/bin/codecept run acceptance,unit,integration,webapi --html -vvv"'
                 }
             }
         }
@@ -120,8 +125,9 @@ pipeline {
                 // export acceptance test report files
                 sh 'cd $WORKSPACE/.docker && docker cp $(docker-compose ps -q bo):/backoffice/tests/_output/ $WORKSPACE/acceptance_test_reports/'
 
-                // export phplint report
+                // export phplint / phpstan report
                 sh 'cd $WORKSPACE/.docker && docker cp $(docker-compose ps -q bo):/backoffice/phplintreport.txt $WORKSPACE/phplintreport.txt'
+                sh 'cd $WORKSPACE/.docker && docker cp $(docker-compose ps -q bo):/backoffice/phpstanreport.txt $WORKSPACE/phpstanreport.txt'
 
                 publishHTML([
                     allowMissing: true,
@@ -141,6 +147,15 @@ pipeline {
                     reportName: 'Phplint - Syntax Error Reports',
                     reportTitles: 'Phplint - Syntax Error Reports'])
 
+                publishHTML([
+                    allowMissing: true,
+                    alwaysLinkToLastBuild: true,
+                    keepAll: true,
+                    reportDir: '',
+                    reportFiles: 'phpstanreport.txt',
+                    reportName: 'Phpstan - Syntax Error Reports',
+                    reportTitles: 'Phpstan - Syntax Error Reports'])
+
                 //sh 'docker cp ac66_bo:/backoffice/tests/_output/coverage/ ./coverage/'
                 //coverage reports
                 //publishHTML([
@@ -152,10 +167,17 @@ pipeline {
                 //    reportName: 'Coverage Report',
                  //   reportTitles: 'Code Coverage Report'])
 
+                script {
+                    if ("${params.DESTROY_CONTAINERS}" == "Yes") {
+                        echo "removing containers"
+                        sh 'cd $WORKSPACE/.docker && (docker-compose down --volumes || true )'
+                    }
+                }
+
                 // cleanup
-                sh 'cd $WORKSPACE/.docker && (docker-compose down --volumes || true )'
+
 		
-		slackSend channel: '#ac66', message: "Testing Result: ${currentBuild.currentResult} for ${GIT_BRANCH}@${GIT_COMMIT} see ${env.BUILD_URL} "
+		        slackSend channel: '#ac66', message: "Testing Result: ${currentBuild.currentResult} for ${GIT_BRANCH}@${GIT_COMMIT} see ${env.BUILD_URL} "
             }
         }
     }
